Hide Routes, unify Router interface, simplify Handlers
If you want to have a route's handler be a pure function of the request, use a
FunctionalHandler. For convenience, you can define a FunctionalHandler with an
integer parameter; this produces a handler that ignores data in a request, and
responds with a response of that status code.

Most handlers, however, will involve some IO. In this case, implement the
Handler interface. Currently undecided: will it make sense to create an
inheritance tree for any of these IO handlers?

Rename 'apply' to 'respond'

CHKPT: Changing RequestHandlers to Handlers wherever possible

Replace RequestHandler with Handler where possible

Remove duplication

Rename and simplify

Eliminate Routes parameter to new Routers

Make Routes package-private

Remove incorrectly placed responsibility; improve name