Add directory structure; name changes
Handler => Responder
A Responder is a thing that responds to Requests with Responses.

Service extends Handler
A Service is a thing that Responds, but it's expected to respond to several
things.

Many Handlers => Services

++

I'm not totally sure that I'll extend the Service abstraction further. But I felt
I needed a name to keep myself straight on the difference between "a thing
that takes a Request and yields a Response" versus "a collection of
Request->Response actions that may interdepend," e.g., the cookie
story in cob_spec.