Handle IOExceptions in Responders with 500 status
It seemed very clear where to create the response with a 500 status once
the compiler explained how IOExceptions get thrown and must eventually be
caught. Learning to raise exceptions myself is not on the critical path
for this iteration, and so this is not tested. I will add that to the
backlog.